fix: replace ShareMenu with ShareButton
